Various fixes.

+++ classes/graphics/backends/gdi/bitmapgdi.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -135,6 +135,8 @@ S::Bool S::GUI::BitmapGDI::DeleteBitmap()
 
 S::Bool S::GUI::BitmapGDI::SetSystemBitmap(Void *nBitmap)
 {
+	if (nBitmap == GetSystemBitmap()) return True;
+
 	if (nBitmap == NIL)
 	{
 		DeleteBitmap();
